+++ src/detections/configs.rs
@@ -5,7 +5,7 @@ use crate::options::geoip_search::GeoIPSearch;
 use crate::options::htmlreport;
 use crate::options::profile::{load_profile, Profile};
 use chrono::{DateTime, Utc};
-use clap::{Args, ColorChoice, Command, CommandFactory, Parser, Subcommand};
+use clap::{ArgGroup, Args, ColorChoice, Command, CommandFactory, Parser, Subcommand};
 use compact_str::CompactString;
 use hashbrown::{HashMap, HashSet};
 use lazy_static::lazy_static;
@@ -18,7 +18,7 @@ use std::{fs, process};
 use terminal_size::{terminal_size, Width};
 use yaml_rust::{Yaml, YamlLoader};
 
-use super::message::create_output_filter_config;
+use super::message::{create_output_filter_config, LEVEL_ABBR_MAP};
 use super::utils::check_setting_path;
 
 lazy_static! {
@@ -98,6 +98,54 @@ impl StoredStatic {
             Some(Action::JsonTimeline(opt)) => opt.json_input,
             _ => false,
         };
+        let is_valid_min_level = match &input_config.as_ref().unwrap().action {
+            Some(Action::CsvTimeline(opt)) => LEVEL_ABBR_MAP
+                .keys()
+                .any(|level| &opt.output_options.min_level.to_lowercase() == level),
+            Some(Action::JsonTimeline(opt)) => LEVEL_ABBR_MAP
+                .keys()
+                .any(|level| &opt.output_options.min_level.to_lowercase() == level),
+            Some(Action::PivotKeywordsList(opt)) => LEVEL_ABBR_MAP
+                .keys()
+                .any(|level| &opt.min_level.to_lowercase() == level),
+            _ => true,
+        };
+        let is_valid_exact_level = match &input_config.as_ref().unwrap().action {
+            Some(Action::CsvTimeline(opt)) => {
+                opt.output_options.exact_level.is_none()
+                    || LEVEL_ABBR_MAP.keys().any(|level| {
+                        &opt.output_options
+                            .exact_level
+                            .as_ref()
+                            .unwrap()
+                            .to_lowercase()
+                            == level
+                    })
+            }
+            Some(Action::JsonTimeline(opt)) => {
+                opt.output_options.exact_level.is_none()
+                    || LEVEL_ABBR_MAP.keys().any(|level| {
+                        &opt.output_options
+                            .exact_level
+                            .as_ref()
+                            .unwrap()
+                            .to_lowercase()
+                            == level
+                    })
+            }
+            Some(Action::PivotKeywordsList(opt)) => {
+                opt.exact_level.is_none()
+                    || LEVEL_ABBR_MAP
+                        .keys()
+                        .any(|level| &opt.exact_level.as_ref().unwrap().to_lowercase() == level)
+            }
+            _ => true,
+        };
+        if !is_valid_min_level || !is_valid_exact_level {
+            AlertMessage::alert(" You specified an invalid level. Please specify informational, low, medium, high or critical.").ok();
+            process::exit(1);
+        }
+
         let geo_ip_db_result = match &input_config.as_ref().unwrap().action {
             Some(Action::CsvTimeline(opt)) => GeoIPSearch::check_exist_geo_ip_files(
                 &opt.geo_ip,
@@ -508,6 +556,7 @@ pub struct MetricsOption {
 }
 
 #[derive(Args, Clone, Debug)]
+#[clap(group(ArgGroup::new("level_rule_filtering").args(["min_level", "exact_level"]).multiple(false)))]
 pub struct PivotKeywordOption {
     #[clap(flatten)]
     pub input_args: InputOption,
@@ -534,6 +583,15 @@ pub struct PivotKeywordOption {
         value_name = "LEVEL"
     )]
     pub min_level: String,
+
+    /// Scan for only specific levels (informational, low, medium, high, critical)
+    #[arg(
+        help_heading = Some("Filtering"),
+        short = 'e',
+        long = "exact-level",
+        value_name = "LEVEL"
+    )]
+    pub exact_level: Option<String>,
 
     /// Enable rules marked as noisy
     #[arg(help_heading = Some("Filtering"), short = 'n', long = "enable-noisy-rules")]
@@ -564,6 +622,7 @@ pub struct LogonSummaryOption {
 
 /// Options can be set when outputting
 #[derive(Args, Clone, Debug)]
+#[clap(group(ArgGroup::new("level_rule_filtering").args(["min_level", "exact_level"]).multiple(false)))]
 pub struct OutputOption {
     #[clap(flatten)]
     pub input_args: InputOption,
@@ -594,6 +653,15 @@ pub struct OutputOption {
         value_name = "LEVEL"
     )]
     pub min_level: String,
+
+    /// Scan for only specific levels (informational, low, medium, high, critical)
+    #[arg(
+        help_heading = Some("Filtering"),
+        short = 'e',
+        long = "exact-level",
+        value_name = "LEVEL"
+    )]
+    pub exact_level: Option<String>,
 
     /// Enable rules marked as noisy
     #[arg(help_heading = Some("Filtering"), short = 'n', long = "enable-noisy-rules")]
@@ -1041,6 +1109,7 @@ fn extract_output_options(config: &Config) -> Option<OutputOption> {
             profile: None,
             exclude_status: option.exclude_status.clone(),
             min_level: option.min_level.clone(),
+            exact_level: option.exact_level.clone(),
             end_timeline: option.end_timeline.clone(),
             start_timeline: option.start_timeline.clone(),
             eid_filter: option.eid_filter,
@@ -1064,6 +1133,7 @@ fn extract_output_options(config: &Config) -> Option<OutputOption> {
             profile: None,
             exclude_status: None,
             min_level: String::default(),
+            exact_level: None,
             end_timeline: None,
             start_timeline: None,
             eid_filter: false,
@@ -1087,6 +1157,7 @@ fn extract_output_options(config: &Config) -> Option<OutputOption> {
             profile: None,
             exclude_status: None,
             min_level: String::default(),
+            exact_level: None,
             end_timeline: None,
             start_timeline: None,
             eid_filter: false,
@@ -1119,6 +1190,7 @@ fn extract_output_options(config: &Config) -> Option<OutputOption> {
             profile: None,
             exclude_status: None,
             min_level: String::default(),
+            exact_level: None,
             end_timeline: None,
             start_timeline: None,
             eid_filter: false,