+++ src/options/update.rs
@@ -197,6 +197,7 @@ impl Update {
             .read_dir(
                 rule_folder_path,
                 "INFORMATIONAL",
+                "",
                 &filter::RuleExclude::new(),
                 stored_staic,
             )