+++ src/yaml.rs
@@ -66,7 +66,8 @@ impl ParseYaml {
     pub fn read_dir<P: AsRef<Path>>(
         &mut self,
         path: P,
-        level: &str,
+        min_level: &str,
+        target_level: &str,
         exclude_ids: &RuleExclude,
         stored_static: &StoredStatic,
     ) -> io::Result<String> {
@@ -152,7 +153,13 @@ impl ParseYaml {
                 let entry = entry?;
                 // フォルダは再帰的に呼び出す。
                 if entry.file_type()?.is_dir() {
-                    self.read_dir(entry.path(), level, exclude_ids, stored_static)?;
+                    self.read_dir(
+                        entry.path(),
+                        min_level,
+                        target_level,
+                        exclude_ids,
+                        stored_static,
+                    )?;
                     return io::Result::Ok(ret);
                 }
                 // ファイル以外は無視
@@ -308,10 +315,14 @@ impl ParseYaml {
                 .unwrap_or("informational")
                 .to_uppercase();
             let doc_level_num = self.level_map.get(doc_level).unwrap_or(&1);
-            let args_level_num = self.level_map.get(level).unwrap_or(&1);
-            if doc_level_num < args_level_num {
+            let args_level_num = self.level_map.get(min_level).unwrap_or(&1);
+            let target_level_num = self.level_map.get(target_level).unwrap_or(&0);
+            if doc_level_num < args_level_num
+                || (target_level_num != &0_u128 && doc_level_num != target_level_num)
+            {
                 return Option::None;
             }
+
             Option::Some((filepath, yaml_doc))
         });
         self.files.extend(files);
@@ -354,6 +365,7 @@ mod tests {
                     enable_deprecated_rules: false,
                     exclude_status: None,
                     min_level: "informational".to_string(),
+                    exact_level: None,
                     enable_noisy_rules: false,
                     end_timeline: None,
                     start_timeline: None,
@@ -387,6 +399,7 @@ mod tests {
         let _ = &yaml.read_dir(
             "test_files/rules/yaml/1.yml",
             &String::default(),
+            "",
             &exclude_ids,
             &dummy_stored_static,
         );
@@ -403,6 +416,7 @@ mod tests {
         let _ = &yaml.read_dir(
             "test_files/rules/yaml/",
             &String::default(),
+            "",
             &exclude_ids,
             &dummy_stored_static,
         );
@@ -444,6 +458,7 @@ mod tests {
         yaml.read_dir(
             path,
             "",
+            "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
         )
@@ -458,7 +473,8 @@ mod tests {
         let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
         yaml.read_dir(
             path,
-            "informational",
+            "INFORMATIONAL",
+            "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
         )
@@ -473,6 +489,7 @@ mod tests {
         yaml.read_dir(
             path,
             "LOW",
+            "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
         )
@@ -487,6 +504,7 @@ mod tests {
         yaml.read_dir(
             path,
             "MEDIUM",
+            "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
         )
@@ -501,6 +519,7 @@ mod tests {
         yaml.read_dir(
             path,
             "HIGH",
+            "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
         )
@@ -515,6 +534,7 @@ mod tests {
         yaml.read_dir(
             path,
             "CRITICAL",
+            "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
         )
@@ -528,6 +548,7 @@ mod tests {
         let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
         yaml.read_dir(
             path,
+            "",
             "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
@@ -543,6 +564,7 @@ mod tests {
         yaml.read_dir(
             path,
             "",
+            "",
             &filter::exclude_ids(&dummy_stored_static),
             &dummy_stored_static,
         )
@@ -555,7 +577,7 @@ mod tests {
         let dummy_stored_static = create_dummy_stored_static();
         let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
         let exclude_ids = RuleExclude::new();
-        yaml.read_dir(path, "", &exclude_ids, &dummy_stored_static)
+        yaml.read_dir(path, "", "", &exclude_ids, &dummy_stored_static)
             .unwrap();
         assert_eq!(yaml.rule_load_cnt.get("excluded").unwrap().to_owned(), 0);
     }
@@ -565,11 +587,91 @@ mod tests {
         let dummy_stored_static = create_dummy_stored_static();
         let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
         let exclude_ids = RuleExclude::new();
-        yaml.read_dir(path, "", &exclude_ids, &dummy_stored_static)
+        yaml.read_dir(path, "", "", &exclude_ids, &dummy_stored_static)
             .unwrap();
         assert_eq!(
             yaml.rule_status_cnt.get("deprecated").unwrap().to_owned(),
             1
         );
+    }
+
+    #[test]
+    fn test_info_exact_level_read_yaml() {
+        let dummy_stored_static = create_dummy_stored_static();
+        let path = Path::new("test_files/rules/level_yaml");
+        let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
+        yaml.read_dir(
+            path,
+            "",
+            "INFORMATIONAL",
+            &filter::exclude_ids(&dummy_stored_static),
+            &dummy_stored_static,
+        )
+        .unwrap();
+        assert_eq!(yaml.files.len(), 1);
+    }
+
+    #[test]
+    fn test_low_exact_level_read_yaml() {
+        let path = Path::new("test_files/rules/level_yaml");
+        let dummy_stored_static = create_dummy_stored_static();
+        let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
+        yaml.read_dir(
+            path,
+            "",
+            "LOW",
+            &filter::exclude_ids(&dummy_stored_static),
+            &dummy_stored_static,
+        )
+        .unwrap();
+        assert_eq!(yaml.files.len(), 1);
+    }
+
+    #[test]
+    fn test_medium_exact_level_read_yaml() {
+        let path = Path::new("test_files/rules/level_yaml");
+        let dummy_stored_static = create_dummy_stored_static();
+        let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
+        yaml.read_dir(
+            path,
+            "",
+            "MEDIUM",
+            &filter::exclude_ids(&dummy_stored_static),
+            &dummy_stored_static,
+        )
+        .unwrap();
+        assert_eq!(yaml.files.len(), 1);
+    }
+
+    #[test]
+    fn test_high_exact_level_read_yaml() {
+        let path = Path::new("test_files/rules/level_yaml");
+        let dummy_stored_static = create_dummy_stored_static();
+        let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
+        yaml.read_dir(
+            path,
+            "",
+            "HIGH",
+            &filter::exclude_ids(&dummy_stored_static),
+            &dummy_stored_static,
+        )
+        .unwrap();
+        assert_eq!(yaml.files.len(), 1);
+    }
+
+    #[test]
+    fn test_critical_exact_level_read_yaml() {
+        let path = Path::new("test_files/rules/level_yaml");
+        let dummy_stored_static = create_dummy_stored_static();
+        let mut yaml = yaml::ParseYaml::new(&dummy_stored_static);
+        yaml.read_dir(
+            path,
+            "",
+            "CRITICAL",
+            &filter::exclude_ids(&dummy_stored_static),
+            &dummy_stored_static,
+        )
+        .unwrap();
+        assert_eq!(yaml.files.len(), 1);
     }
 }